==> part1_data_cleaning.Rmd <==
---
title: "BST 260 Final Project: What should we care about? Contraceptive Use and Health!"
author: "Binkai (Cathy) Liu, Karla Flores Guzman, Xinhui (Kiera) Zhang, Yichi Zhang, Jiabao (Lily) Zhong"
date: "12/12/2021"
output: html_document
---

Note: All original datasets used for our analysis are accessible through both our github repository and website links in later sections. You can also direct to the catelog in the Readme.md file in our github repository to check data sources.

## Overview and Motivation
In 2019, over 1 out of 1.9 billion Women of the Reproductive Age group (15-49 years old) worldwide need family planning, according to WHO; of these, around 800 million use contraceptive methods. Unfortunately, almost 300 million have an unmet need for contraception (WHO). Access to contraceptives is heterogeneous around the world. There could be many reasons for the variation in contraceptives use: taboos, lack of sexual education, armed conflicts, to name a few. Also, women’s acceptance could be diminished by religious opposition, fear of side effects, and poor quality of medical counseling and service.
It is a human right to decide the number and spacing of their children; contraceptives play a crucial role in helping to achieve that right. Moreover, it is well documented that the use of condoms helps to reduce the risk of getting sexually transmitted diseases.
Our team comprises five women interested in different aspects of human health using epidemiological approaches. Motivated by our identities as women scholars from different parts of the world, we were inspired to explore more about how contraceptive uses are associated with different health outcomes across countries.  

Reference: [WHO. Family planning/contraception methods](https://www.who.int/news-room/fact-sheets/detail/family-planning-contraception)

## Related Work
During the time when we were deciding topics to choose for our project, we brought up recent discussions about [abortion law in Texas](https://www.texastribune.org/2021/10/29/texas-abortion-law-supreme-court/). As women scholars from different regions of the world, we thought it would be interesting to look into how abortion law in different countries affects reproductive health worldwide. As we explored more into different dataset available online, we specified our outcomes to be birth rate, infant mortality rate, and HIV prevalence, and decided to explore associations between various socioeconomic factors and outcomes through statistical models. 

## Initial Questions
Our study objectives are assessing associations between contraceptive use and our outcomes (birth rate, infant mortality rate, and HIV prevalence), respectively, using linear regression, K-nearest neighbors (kNN), and decision tree models. We also present an interactive surface through Shiny App to display different contraceptives uses in different countries in the world, and to let our readers explore the associations between contraceptive use and health together with us.

## Data
A big strength of our project is rich data from different credible sources, including World Health Organization, other departments of the United Nations, and Gapminder. Each group member was responsible for at least some parts of data obtaining, cleaning, and compiling. In the following sections, we will be loading in dataset we used, and presenting data source, brief description of the dataset, and data wrangling we conducted on each dataset obtained.


### Load packages
```{r load packages, message=FALSE, warning=FALSE}
library(tidyverse) 
library(readr)
library(dplyr)
library(rvest)
library(stringr)
library(readxl)
library(psych)
```

### Read in data sets & perform preliminary data wrangling

#### *Predictor: Geographic region*

* Data source: [United Nations, Department of Economic and Social Affairs
Population Dynamics, World Population Prospects 2019, Births.xslx](https://population.un.org/wpp/Download/Standard/Fertility/)
* Data description: We adopted official classification of countries and geographic regions based on the birth rate dataset downloaded from the United Nations website. There are 8 regions in total. This dataset needs to be loaded and cleaned first because we need to incorporate geographic region data into other dataset loaded later.

```{r geographic region}
# code geographic regions & matched them with parent code in the UN data sets
geographic <- data.frame(region_l = c(rep("Sub-saharan Africa", 4),
                                      rep("Northern Africa and Western Asia", 2),
                                      rep("Central and Southern Asia", 2),
                                      rep("Eastern and South-eastern Asia", 2),
                                      rep("Latin America and the Caribbean", 3),
                                      "Australia and New Zealand",
                                      rep("Other Oceania", 3),
                                      rep("Europe and Northern America", 5)),
                         code = c(910, 911, 913, 914, 912, 922, 5500, 5501, 906, 
                                  920, 915, 916, 931, 927, 928, 954, 957, 923, 
                                  924, 925, 926, 918))

# dummy code region_l (categorical var.) 
regionL <- as.data.frame(dummy.code(geographic$region_l))

# rename the dummy coded variables & combine them with the original data set
regionL <- rename(regionL, r_Subsaharan_Africa = `Sub-saharan Africa`)
regionL <- rename(regionL, r_Euro_N.America = `Europe and Northern America`)
regionL <- rename(regionL, r_LatinAmerica = `Latin America and the Caribbean`)
regionL <- rename(regionL, r_NW.Africa = `Northern Africa and Western Asia`)
regionL <- rename(regionL, r_ESE.Asia = `Eastern and South-eastern Asia`)
regionL <- rename(regionL, r_CS.Asia = `Central and Southern Asia`)
regionL <- rename(regionL, r_OtherOceania = `Other Oceania`)
regionL <- rename(regionL, r_AU_NZ = `Australia and New Zealand`)
geographic <- cbind(geographic, regionL)
```


#### *Outcome 1: Infant mortality rate*

* Data source: [World Health Organization, The Global Health Observatory, Infant mortality rate (probability of dying between birth and age 1 per 1000 live births)](https://www.who.int/data/gho/data/indicators/indicator-details/GHO/infant-mortality-rate-(probability-of-dying-between-birth-and-age-1-per-1000-live-births))
* Data description: This dataset from WHO presents infant mortality rate (probability of dying between birth and age 1 per 1000 live births) in each country. We used the most recent year data for all countries (up to 2019).

```{r infant mortality rate, message=FALSE}
# read in the data set
infant <- read_csv("rawdata_infant_mortality.csv")

# select the columns we want
infant <- infant %>% 
  filter(IsLatestYear=="TRUE", Dim1 == "Both sexes") %>% 
  dplyr::select(`Location`,`Period`, `FactValueNumeric`)
```

#### *Outcome 2: Crude birth rate*

* Data source: [United Nations, Department of Economic and Social Affairs
Population Dynamics, World Population Prospects 2019, Births.xslx](https://population.un.org/wpp/Download/Standard/Fertility/)
* Data description: This dataset presents number of births over 5 years in each country. We used the crude birth rate of 2015-2020 for each country. The 5-year period refers to 1 July, 2015 to 30 June, 2020. Data are presented in thousands.

```{r crude birth rate}
# read in the data set
BR <- read_xlsx("rawdata_birth_rate.xlsx", range = "ESTIMATES!C17:U272")

# data wrangling
BR <- BR %>% 
  dplyr::select(`Region, subregion, country or area *`, `Parent code`, `2015-2020`) %>% 
  # keep relevant info
  setNames(c("country_area", "code", "crude_birth_rate_perK")) %>% 
  # rename: no spaces in names
  filter(code %in% geographic$code) %>% 
  # remove rows for world/regions/etc., keep country-level info
  full_join(geographic, by = "code") %>% 
  # add geographic region to the data set
  dplyr::select(-code) %>% 
  # remove variable "code"
  mutate(crude_birth_rate_perK = as.numeric(crude_birth_rate_perK)) 
  # change parameter to numeric
```

#### *Outcome 3: HIV prevalence*

* Data source: [Gapminder](https://www.gapminder.org/data/)  Path: Health → HIV → Prevalence of HIV among adults age 15-49 
* Data description: This dataset presents the estimated percentage of adults aged 15-49 that are affected by HIV, including those without symptoms, those sick from AIDS, and those healthy due to treatment of the HIV infection. We used the most recent year data (up to 2010), and excluded countries with only data before 2020.

```{r HIV, message=FALSE}
# read in the data set
HIV_prev_raw<- read_csv("rawdata_HIV_prevalence.csv")

# convert to long format & remove NAs
HIV_prev_long <- HIV_prev_raw %>%
  gather( year, HIV_prev, `1978`:`2010`) %>%
  drop_na()

# keep the most recent data
HIV_prev_latest <- HIV_prev_long %>% group_by(country) %>% slice_tail()
  # all data comes after 2000

# change year column name
colnames(HIV_prev_latest)[2]<-"HIV_prev_latest_year"

# delete unwanted data sets
rm(HIV_prev_raw); rm(HIV_prev_long)
```



#### *Predictor: GDP*

* Data source: [Gapminder](https://www.gapminder.org/data/) Path: Economy→Income & Growth → GDP/Capita (US$, inflation adjusted)
* Data description: GDP per capita (divided by midyear population), calculated in constant 2010 US dollars. We used the most recent year (up to 2019) data for all countries.

```{r predictor GDP, warning=FALSE, message=FALSE}
# read in the data set
gdp_raw <- read_csv("rawdata_gdp.csv") # wide format

# convert to long format & remove NAs
gdp_long <- gdp_raw %>% gather(year, gdp, `1959`:`2019`) %>% drop_na()

# keep the most recent data
gdp_latest <- gdp_long %>% group_by(country) %>% slice_tail() %>% ungroup()

# convert "k" into thousand
gdp_latest <- gdp_latest %>%
  mutate(gdp = case_when(
    str_detect(gdp, "k") ~ as.numeric(str_extract(gdp, "[\\d\\.]+")) * 1000,
    TRUE ~ as.numeric(gdp)))

# change year column name
colnames(gdp_latest)[2]<-"gdp_latest_year"

# delete unwanted data sets
rm(gdp_raw); rm(gdp_long)
```

#### *Predictor: Education level*

* Data source: [Gapminder](https://www.gapminder.org/data/)  Path: Education → Mean years in school → % people 15 years or older
* Data description: Average years in school for women of reproductive age 15 to 44, including primary, secondary and tertiary education. Since gapminder did not have a data for both sexes and our interest is reproductive health, we chose female gender group for this predictor. We used the most recent year data (up to 2010) for all countries.

```{r education level, message=FALSE}
# read in the data set
edu_raw <-read_csv("rawdata_edu_level.csv")

# convert to long format & remove NAs
edu_long <- edu_raw %>% 
  gather(year, years_in_school, `1969`:`2008`) %>%
  drop_na()

# keep the most recent data
edu_latest <- edu_long %>% group_by(country) %>% slice_tail()

# change year column name
colnames(edu_latest)[2]<-"edu_latest_year"

# delete unwanted data sets
rm(edu_raw); rm(edu_long)
```



#### *Predictor: Contraceptive use*

* Data source: [United Nations, Population Division, World Contraceptive Use 2021](https://www.un.org/development/desa/pd/data/world-contraceptive-use)
* Data description: According to UN, the World Contraceptive Use 2021, includes country-specific survey-based observations of key family planning indicators, based on survey data available as of January 2021. We used the data of the most recent year for each country, and excluded countries with the most recent year before 2000. Contraceptive use prevalence is used for model building. We also display prevalence of different subtypes of contraceptives use in different countries in world map using Shiny App.

```{r contraceptives, message=FALSE}
# read in the data set
CM <- read_xlsx("rawdata_contraceptives.xlsx", 
                range = "By methods!A8:W1377", na = "..")

# set names
names(CM) <- as.character(c("country_area", "ISO", "start", "end", "age_group", 
                            CM[1, 6:23]))

# clean up variable names
names(CM) <- names(CM) %>%
  str_to_lower() %>%
  str_replace_all("[\\n]", "") %>% # remove "\n" in female/male sterilization
  str_replace_all("\\s\\(\\w+\\)$", "") %>% # remove " (LAM)"
  str_replace_all("[\\s\\/]", "_") # replace spaces with underscores

# remove first row with only column name information
CM <- CM[-1,] %>%  
  mutate_at(6:23, as.numeric) # change to numeric

# write a function to get the most recent usage% (after 2000) of a given method
most_recent <- function(var){ 
  CM %>%
    filter(end >= 2000) %>% drop_na(!!var) %>% group_by(country_area) %>% 
    slice(which.max(end)) %>% dplyr::select(country_area, !!var)}
  # function input should be enclosed in quotes

# create a variable to store contraceptive methods
methods <- names(CM)[6:23] 

# create a new data frame; first column = 196 distinct country names
CM_2000 <- data.frame(country_area = levels(as.factor(CM$country_area)))

# loop over all contraceptive methods & use them as function input
for (method in methods){ 
  temp <- as.data.frame(most_recent(method)) 
  CM_2000 <- left_join(CM_2000, temp, by = "country_area")
    # append the outputs to CM_2000
  rm(temp)}

# collapse some categories of contraceptive methods into larger groups
CM_2000 <- CM_2000 %>% 
  mutate(invasive = rowSums(.[,c("female_sterilization", "male_sterilization", 
                                 "iud", "implant")], na.rm = T),
         medication = rowSums(.[,c("injectable", "pill")], na.rm = T),
         barrier = rowSums(.[,c("male_condom", "female_condom", 
                                "vaginal_barrier_methods")], na.rm = T),
         other_modern = rowSums(.[,c("lactational_amenorrhea_method", 
                                     "emergency_contraception", 
                                     "other_modern_methods")], na.rm = T),
         traditional = rowSums(.[,c("rhythm", "withdrawal", 
                                    "other_traditional_methods")], na.rm = T)) 

# remove unwanted columns
CM_2000 <- CM_2000 %>% 
  dplyr::select(c(country_area, any_method, any_modern_method,
                  any_traditional_method, invasive, medication, barrier, 
                  other_modern, traditional))

# delete unwanted objects
rm(CM); rm(method); rm(methods); rm(most_recent)
```

#### *Predictor: Mean age of childbearing*

* Data source: [United Nations, Department of Economic and Social Affairs
Population Dynamics, World Population Prospects 2019, Mean Age of Childbearing.xslx](https://population.un.org/wpp/Download/Standard/Fertility/)
* Data description: According to UN, this dataset represents the average age of mothers at the birth of their children if women were subject throughout their lives to the age-specific fertility rates observed in a given year. We used data of year `2015-2020` for all countries.

```{r mean age of childbearing}
# read in the data set
MAC <- read_xlsx("rawdata_mean_age_childbearing.xlsx", 
                 range = "ESTIMATES!C17:U272")

# data wrangling
MAC <- MAC %>% 
  dplyr::select(`Region, subregion, country or area *`, `Parent code`, `2015-2020`) %>% 
  # keep relevant info
  setNames(c("country_area", "code", "mean_age_of_childbearing")) %>% 
  # rename: no spaces in names
  filter(code %in% geographic$code) %>% 
  # remove rows for world/regions/etc., keep country-level info
  dplyr::select(-code) %>% 
  # remove variable "code"
  mutate(mean_age_of_childbearing = as.numeric(mean_age_of_childbearing))
```

#### *Predictor: Abortion law score*

* Data source: [Wikipedia: Abortion Law. Table. Legal grounds on which abortion is permitted in independent countries](https://en.wikipedia.org/wiki/Abortion_law)
* Data description: This dataset was "web-scrapped" from Wikipedia (Contributer: Binkai (Cathy) Liu), which summarizes the legal grounds for abortion in all United Nations member states and United Nations General Assembly observer states and some countries with limited recognition (202 countries/regions). The data from Wikipedia is mostly based on data compiled by the United Nations up to 2019. We cleaned the dataset and further calculated `Abortion law score` for each country based on the numbers of times that abortion being prohibited in the 6 situations presented in the table, including "Risk to life, risk to health, rape, fetal impairment, economic or social, and on request". The `Abortion law score` ranges from 0-6, with 6 being the most strict on abortion law.

```{r abortion law, warning = FALSE, message= FALSE}
# web scrap the abortion law table by country
url <- 'https://en.wikipedia.org/wiki/Abortion_law'
h <- read_html(url)
tab <- h %>% html_nodes("table") %>% .[6] %>% html_table %>% .[[1]] 

# remove rows that are subdivisions of a country
new_tab <- tab[-c(12:23, 38:40, 50:62, 148:179, 197:233, 299:302, 304:354),]
  # Australia (row 12-23); Bosnia and Herzegovina (row 38-40)
  # Canada (row 50-62); Mexico (row 148-179); Nigeria (row 197-233)
  # United kingdom (row 299-302); United states (row 304-354)
  # current rows = 212

# remove rows with "Country" in the country column (subtitle rows)
new_tab <- subset(new_tab, Country != "Country")
  # current rows = 202

# delete citation number of country name 
test <- gsub("(\\[).*", "", new_tab$Country) %>% str_trim(.)
clean_tab <- new_tab %>% mutate(Country = test)

# assign rigorous score of abortion law by country:
  # how many "prohibited" there are in each row (for each country)
score <- mutate(clean_tab, score = 
          str_count(clean_tab$`Risk to life`, "prohibited")
         +str_count(clean_tab$`Risk to health`, "prohibited")
         +str_count(clean_tab$Rape, "prohibited")
         +str_count(clean_tab$`Fetal impairment`, "prohibited")
         +str_count(clean_tab$`Economic or social`, "prohibited")
         +str_count(clean_tab$`On request`, "prohibited"))
score <- score %>% dplyr::select(c("Country", "score"))
names(score) <- c("country", "score")

# delete unwanted data sets & vectors
rm(url); rm(h); rm(tab); rm(new_tab); rm(clean_tab); rm(test)
```



### Join the tables & Final clean-up

```{r join the tables, warning=FALSE, message=FALSE}
#### combime data sets from the same source ####

# combine gapminder data sets
gapminder_predictors <- full_join(gdp_latest, edu_latest)
gapminder_full <- full_join(gapminder_predictors, HIV_prev_latest)

# combine UN data sets
CM_2000$country_area <- 
  str_replace_all(CM_2000$country_area, "Democratic People's Republic of Korea",
                  "Dem. People's Republic of Korea")
BR$country_area <- 
  str_replace_all(BR$country_area, "Czechia", "Czech Republic")
MAC$country_area <- 
  str_replace_all(MAC$country_area, "Czechia", "Czech Republic")

UNdata_BR_CM_MAC <- full_join(BR,CM_2000) %>% full_join(.,MAC)
colnames(UNdata_BR_CM_MAC)[1] <- "country"

# combine WHO data sets
colnames(infant)[3]<- "infant_mortality"
WHOdata_infant <- infant
colnames(WHOdata_infant)[1] <- "country"


#### recode country names (reference: abortion law table) ####

# recode country names in gapminder data sets
country_key_gap = c(
  "Congo, Rep."="Congo",
  "Congo, Dem. Rep."="Democratic Republic of the Congo", 
  "Timor-Leste"="East Timor",
  "Cote d'Ivoire"="Ivory Coast",
  "Kyrgyz Republic"="Kyrgyzstan",
  "Lao"="Laos",
  "Micronesia, Fed. Sts."="Micronesia",
  "St. Kitts and Nevis"="Saint Kitts and Nevis",
  "St. Lucia"="Saint Lucia",
  "St. Vincent and the Grenadines"="Saint Vincent and the Grenadines",
  "Sao Tome and Principe"="São Tomé and Príncipe",
  "Slovak Republic"="Slovakia",
  "American Samoa"="Samoa",
  "Hong Kong, China"="China, Hong Kong SAR",
  "Macao, China"="China, Macao SAR",
  "Virgin Islands (U.S.)"="United States Virgin Islands")
gapminder_full <- gapminder_full %>% 
  mutate(country = recode_factor(country, !!!country_key_gap))

# recode country names in UN data sets
country_key_UN = c(
  "Bolivia (Plurinational State of)"="Bolivia",
  "Brunei Darussalam"="Brunei", 
  "Cabo Verde"="Cape Verde",
  "Timor-Leste"="East Timor",
  "Iran (Islamic Republic of)"="Iran",
  "Côte d'Ivoire"="Ivory Coast",
  "Lao People's Democratic Republic"="Laos",
  "Micronesia (Fed. States of)"="Micronesia",
  "Republic of Moldova"="Moldova",
  "Dem. People's Republic of Korea"="North Korea",
  "State of Palestine"="Palestine",
  "Russian Federation"="Russia",
  "Sao Tome and Principe"="São Tomé and Príncipe",
  "Republic of Korea"="South Korea",
  "Syrian Arab Republic"="Syria",
  "China, Taiwan Province of China"="Taiwan",
  "United Republic of Tanzania"="Tanzania",
  "United States of America"="United States",
  "Venezuela (Bolivarian Republic of)"="Venezuela",
  "Viet Nam"="Vietnam")
UNdata_BR_CM_MAC <- UNdata_BR_CM_MAC %>%   
  mutate(country = recode_factor(country, !!!country_key_UN))

# recode country names in WHO data sets
country_key_WHO = c(
  "Bolivia (Plurinational State of)"="Bolivia",
  "Brunei Darussalam"="Brunei", 
  "Cabo Verde"="Cape Verde",
  "Czechia"="Czech Republic",
  "Timor-Leste"="East Timor",
  "Iran (Islamic Republic of)"="Iran",
  "Côte d’Ivoire"="Ivory Coast",
  "Lao People's Democratic Republic"="Laos",
  "Micronesia (Federated States of)"="Micronesia",
  "Republic of Moldova"="Moldova",
  "Democratic People's Republic of Korea"="North Korea",
  "The former Yugoslav Republic of Macedonia"="North Macedonia",
  "Russian Federation"="Russia",
  "Sao Tome and Principe"="São Tomé and Príncipe",
  "Republic of Korea"="South Korea",
  "Syrian Arab Republic"="Syria",
  "United Republic of Tanzania"="Tanzania",
  "United Kingdom of Great Britain and Northern Ireland"="United Kingdom",
  "United States of America"="United States",
  "Venezuela (Bolivarian Republic of)"="Venezuela",
  "Viet Nam"="Vietnam")
WHOdata_infant <- WHOdata_infant %>%   
  mutate(country = recode_factor(country, !!!country_key_WHO))


#### combine all data sets ####
Score_GAP_UN_WHO <- full_join(score, gapminder_full) %>%
  full_join(., UNdata_BR_CM_MAC) %>%
  full_join(., WHOdata_infant)
world_clean <- Score_GAP_UN_WHO %>% 
  dplyr::select(-gdp_latest_year, -edu_latest_year, 
         -HIV_prev_latest_year, -Period)

####  remove temporary data sets #### 
rm(list=setdiff(ls(), "world_clean"))

#### export data frame to csv ####
write.csv(world_clean,"world_clean.csv", row.names = FALSE)
```


### Merge cleaned data set (`world_clean.csv`) with maps for Shiny app

```{r}
# load the map package
library(maps)
```

```{r}
# load data sets
clean <- world_clean # the cleaned data set
world <- map_data("world") # world map

# create variables needed for plotting
clean <- clean %>% 
    mutate(prop_modern = any_modern_method/any_method*100) %>% # percentage of modern methods
    mutate(prop_invasive = invasive/any_method*100) %>% # percentage of invasive methods
    mutate(prop_medication = medication/any_method*100) %>% # percentage of medication methods
    mutate(prop_barrier = barrier/any_method*100) %>% # percentage of barrier methods
    mutate(prop_otherM = other_modern/any_method*100) %>% # percentage of other modern methods
    mutate(prop_traditional = traditional/any_method*100) # percentage of traditional methods

# some countries have > 100% modern or traditional methods, convert to 100% 
clean$prop_modern[which(clean$prop_modern > 100)] <- 100
clean$prop_traditional[which(clean$prop_traditional > 100)] <- 100

# compare to find difference in country names for matching
only_clean <- setdiff(clean$country, world$region) 
only_world <- setdiff(world$region, clean$country)
rm(only_clean); rm(only_world)

# match names
match <- data.frame(rbind(c("Antigua and Barbuda", "Antigua"),
                          c("Congo", "Republic of Congo"),
                          c("East Timor", "Timor-Leste"),
                          c("Eswatini", "Swaziland"),
                          c("Saint Kitts and Nevis", "Saint Kitts"),
                          c("Saint Vincent and the Grenadines", "Saint Vincent"),
                          c("São Tomé and Príncipe", "Sao Tome and Principe"),
                          c("Trinidad and Tobago", "Trinidad"),
                          c("United Kingdom", "UK"),
                          c("United States", "USA"),
                          c("Vatican City", "Vatican"),
                          c("Curaçao", "Curacao"),
                          c("Faeroe Islands", "Faroe Islands"),
                          c("Sint Maarten (Dutch part)", "Sint Maarten"),
                          c("United States Virgin Islands", "Virgin Islands"),
                          c("Réunion", "Reunion")))
names(match) <- c("clean", "map")

# character vector with values from map & names from clean
recode <- setNames(match$map, match$clean)

# Recode country names in clean to match with world map
clean <- clean %>% mutate(country = recode(country, !!!recode))

# join clean with world map data set
clean <- clean %>% 
    full_join(world, by = c("country" = "region"))

# exclude columns not relevant for ShinyApp
clean <- clean[, c(1:7, 16:34)]

# write the data set for Shiny to use
write.csv(clean, "shiny_clean.csv", row.names = FALSE)
```
