==> part3_machine_learning.Rmd <==
---
output: html_document
---

## Machine Learning: Decision Tree and KNN Regression

What are the different statistical methods you considered? Justify the decisions you made, and show any major changes to your ideas. How did you reach these conclusions?

### Load packages & Read in the cleaned data set
```{r load packages, message=F, warning = F}
library(tidyverse); library(readr)
library(caret); library(psych)
library(rpart); library(rpart.plot); library(FNN)
library(ggplot2); library(ggthemes); library(ggpubr)

world_clean <- read_csv("world_clean.csv")
df.machineL <- world_clean %>% drop_na()
rm(world_clean)
```

### Decision Trees for Birth Rate and Infant Mortality Rate

#### Create training and testing datasets for decision trees
```{r split data sets - decision tree}
set.seed(260)

# create indices for training set (training:testing = 7:3)
index.train.BR <- 
  createDataPartition(y = df.machineL$crude_birth_rate_perK, p = 0.7, list = F)
index.train.IM <- 
  createDataPartition(y = df.machineL$infant_mortality, p = 0.7, list = F)
index.train.HIV <- 
  createDataPartition(y = df.machineL$HIV_prev, p = 0.7, list = F)

# split the data sets
df.BR.train.tree <- slice(df.machineL, index.train.BR) 
df.BR.test.tree <- slice(df.machineL, -index.train.BR) 
df.IM.train.tree <- slice(df.machineL, index.train.IM) 
df.IM.test.tree <- slice(df.machineL, -index.train.IM) 
df.HIV.train.tree <- slice(df.machineL, index.train.HIV) 
df.HIV.test.tree <- slice(df.machineL, -index.train.HIV) 
```

#### Run decision trees
```{r run decision trees, fig.align="center"}
# fit the model
model.BR.tree <- rpart(crude_birth_rate_perK ~ r_Subsaharan_Africa + 
                         r_Euro_N.America  + r_LatinAmerica + r_NW.Africa + 
                         r_ESE.Asia + r_CS.Asia + r_OtherOceania + r_AU_NZ + 
                         years_in_school + gdp + score + any_method + 
                         mean_age_of_childbearing, 
                  data = df.BR.train.tree)
model.IM.tree <- rpart(HIV_prev ~ r_Subsaharan_Africa + 
                         r_Euro_N.America  + r_LatinAmerica + r_NW.Africa + 
                         r_ESE.Asia + r_CS.Asia + r_OtherOceania + r_AU_NZ + 
                         years_in_school + gdp + score + any_method + 
                         mean_age_of_childbearing, 
                  data = df.IM.train.tree)
model.HIV.tree <- rpart(HIV_prev ~ r_Subsaharan_Africa + 
                         r_Euro_N.America  + r_LatinAmerica + r_NW.Africa + 
                         r_ESE.Asia + r_CS.Asia + r_OtherOceania + r_AU_NZ + 
                         years_in_school + gdp + score + any_method, 
                  data = df.HIV.train.tree)

# prune the tree
mincp.BR <- 
  model.BR.tree$cptable[which.min(model.BR.tree$cptable[,"xerror"]),"CP"]
model.BR.pruned <- 
  prune(model.BR.tree, cp = mincp.BR) 
mincp.IM <- 
  model.IM.tree$cptable[which.min(model.IM.tree$cptable[,"xerror"]),"CP"]
model.IM.pruned <- 
  prune(model.IM.tree, cp = mincp.IM) 
mincp.HIV <- 
  model.HIV.tree$cptable[which.min(model.HIV.tree$cptable[,"xerror"]),"CP"]
model.HIV.pruned <- 
  prune(model.HIV.tree, cp = mincp.HIV) 

# calculate MSEs
pred.BR.train.pruned <- 
  predict(model.BR.pruned, newdata = df.BR.train.tree)
pred.BR.test.pruned <- 
  predict(model.BR.pruned, newdata = df.BR.test.tree)
MSE.BR.train.pruned <- # 15.73
  mean((pred.BR.train.pruned - df.BR.train.tree$crude_birth_rate_perK)^2) 
MSE.BR.test.pruned <- # 24.97
  mean((pred.BR.test.pruned - df.BR.test.tree$crude_birth_rate_perK)^2) 
pred.IM.train.pruned <- 
  predict(model.IM.pruned, newdata = df.IM.train.tree)
pred.IM.test.pruned <- 
  predict(model.IM.pruned, newdata = df.IM.test.tree)
MSE.IM.train.pruned <- # 75.78
  mean((pred.IM.train.pruned - df.IM.train.tree$infant_mortality)^2) 
MSE.IM.test.pruned <- # 94.55
  mean((pred.IM.test.pruned - df.IM.test.tree$infant_mortality)^2) 
pred.HIV.train.pruned <- 
  predict(model.HIV.pruned, newdata = df.HIV.train.tree)
pred.HIV.test.pruned <- 
  predict(model.HIV.pruned, newdata = df.HIV.test.tree)
MSE.HIV.train.pruned <- # 8.51
  mean((pred.HIV.train.pruned - df.HIV.train.tree$HIV_prev)^2) 
MSE.HIV.test.pruned <- # 8.29
  mean((pred.HIV.test.pruned - df.HIV.test.tree$HIV_prev)^2) 

# calculate Pearson coefficients
pearson.BR.train.pruned <- 
  cor.test(pred.BR.train.pruned, df.BR.train.tree$crude_birth_rate_perK,
           method="pearson", conf.level=0.95)$estimate # 0.92 
pearson.BR.test.pruned <- 
  cor.test(pred.BR.test.pruned, df.BR.test.tree$crude_birth_rate_perK,
           method="pearson", conf.level=0.95)$estimate # 0.88
pearson.IM.train.pruned <-
  cor.test(pred.IM.train.pruned, df.IM.train.tree$infant_mortality,
           method="pearson", conf.level=0.95)$estimate # 0.90
pearson.IM.test.pruned <-
  cor.test(pred.IM.test.pruned, df.IM.test.tree$infant_mortality,
           method="pearson", conf.level=0.95)$estimate # 0.90
pearson.HIV.train.pruned <-
  cor.test(pred.HIV.train.pruned, df.HIV.train.tree$HIV_prev,
           method="pearson", conf.level=0.95)$estimate # 0.763
pearson.HIV.test.pruned <-
  cor.test(pred.HIV.test.pruned, df.HIV.test.tree$HIV_prev,
           method="pearson", conf.level=0.95)$estimate # 0.778

# plot the tree
rpart.plot(model.BR.pruned, digits = 4)
rpart.plot(model.IM.pruned, digits = 4)
rpart.plot(model.HIV.pruned, digits = 4)

# delete unwanted objects
rm(mincp.BR); rm(mincp.IM); rm(mincp.HIV)
```


### KNN regression for Birth Rate and Infant Mortality Rate

#### Scale continuous var. for KNN regression
```{r}
# scale the continuous variables
df.machineL.scaled <- df.machineL %>% dplyr::select(-c(7, 17:23))
df.machineL.scaled[, c("years_in_school", "gdp", "score", "any_method", "mean_age_of_childbearing")] <-
  scale(df.machineL.scaled[, c("years_in_school", "gdp", "score", "any_method", "mean_age_of_childbearing")])
```

#### Create training and testing datasets for KNN regression
```{r split data sets - knn}
set.seed(260)

# split the data sets (train set = 97 rows, test set = 40 rows)
df.BR.train.knn <- slice(df.machineL.scaled, index.train.BR) %>% 
  dplyr::select(-infant_mortality, -HIV_prev) 
df.BR.test.knn <- slice(df.machineL.scaled, -index.train.BR) %>% 
  dplyr::select(-infant_mortality, -HIV_prev) 
df.IM.train.knn <- slice(df.machineL.scaled, index.train.IM) %>% 
  dplyr::select(-crude_birth_rate_perK, -HIV_prev) 
df.IM.test.knn <- slice(df.machineL.scaled, -index.train.IM) %>% 
  dplyr::select(-crude_birth_rate_perK, -HIV_prev) 
df.HIV.train.knn <- slice(df.machineL.scaled, index.train.HIV) %>% 
  dplyr::select(-crude_birth_rate_perK, -infant_mortality) 
df.HIV.test.knn <- slice(df.machineL.scaled, -index.train.HIV) %>% 
  dplyr::select(-crude_birth_rate_perK, -infant_mortality) 

# delete unwanted objects
rm(index.train.BR); rm(index.train.IM); rm(index.train.HIV)

# split outcome variable and predictors into two data sets
df.BR.train.knn.X <- df.BR.train.knn %>% 
  dplyr::select(-c(country, crude_birth_rate_perK))
df.BR.train.knn.Y <- df.BR.train.knn %>% 
  dplyr::select(crude_birth_rate_perK)
df.BR.test.knn.X <- df.BR.test.knn %>% 
  dplyr::select(-c(country, crude_birth_rate_perK))
df.BR.test.knn.Y <- df.BR.test.knn %>% 
  dplyr::select(crude_birth_rate_perK)
df.IM.train.knn.X <- df.IM.train.knn %>% 
  dplyr::select(-c(country, infant_mortality))
df.IM.train.knn.Y <- df.IM.train.knn %>% 
  dplyr::select(infant_mortality)
df.IM.test.knn.X <- df.IM.test.knn %>% 
  dplyr::select(-c(country, infant_mortality))
df.IM.test.knn.Y <- df.IM.test.knn %>% 
  dplyr::select(infant_mortality)
df.HIV.train.knn.X <- df.HIV.train.knn %>% 
  dplyr::select(-c(country, HIV_prev, mean_age_of_childbearing))
df.HIV.train.knn.Y <- df.HIV.train.knn %>% 
  dplyr::select(HIV_prev)
df.HIV.test.knn.X <- df.HIV.test.knn %>% 
  dplyr::select(-c(country, HIV_prev, mean_age_of_childbearing))
df.HIV.test.knn.Y <- df.HIV.test.knn %>% 
  dplyr::select(HIV_prev)
```

#### Run knn regression

```{r run knn regression}
# run knn regression on birth rate and infant mortality rate
pred.BR.train.knn <- 
  knn.reg(train = df.BR.train.knn.X, 
          y = df.BR.train.knn.Y$crude_birth_rate_perK, k = 10) %>% .$pred
pred.IM.train.knn <- 
  knn.reg(train = df.IM.train.knn.X, 
          y = df.IM.train.knn.Y$infant_mortality, k = 10) %>% .$pred
pred.HIV.train.knn <- 
  knn.reg(train = df.HIV.train.knn.X, 
          y = df.HIV.train.knn.Y$HIV_prev, k = 10) %>% .$pred
pred.BR.test.knn  <- 
  knn.reg(train = df.BR.train.knn.X, test = df.BR.test.knn.X,
          y = df.BR.train.knn.Y$crude_birth_rate_perK, k = 10) %>% .$pred
pred.IM.test.knn <- 
  knn.reg(train = df.IM.train.knn.X, test = df.IM.test.knn.X, 
          y = df.IM.train.knn.Y$infant_mortality, k = 10) %>% .$pred
pred.HIV.test.knn <- 
  knn.reg(train = df.HIV.train.knn.X, test = df.HIV.test.knn.X, 
          y = df.HIV.train.knn.Y$HIV_prev, k = 10) %>% .$pred

# calculate MSEs
MSE.BR.train.knn <- # 20.58174
  mean((pred.BR.train.knn - df.BR.train.knn.Y$crude_birth_rate_perK)^2)
MSE.BR.test.knn <- # 17.29763
  mean((pred.BR.test.knn - df.BR.test.knn.Y$crude_birth_rate_perK)^2)
MSE.IM.train.knn <- # 124.1427
  mean((pred.IM.train.knn - df.IM.train.knn.Y$infant_mortality)^2)
MSE.IM.test.knn <- # 56.10523
  mean((pred.IM.test.knn - df.IM.test.knn.Y$infant_mortality)^2)
MSE.HIV.train.knn <- # 9.607493
  mean((pred.HIV.train.knn - df.HIV.train.knn.Y$HIV_prev)^2)
MSE.HIV.test.knn <- # 11.44806
  mean((pred.HIV.test.knn - df.HIV.test.knn.Y$HIV_prev)^2)

# calculate Pearson coefficients
pearson.BR.train.knn <- 
  cor.test(pred.BR.train.knn, df.BR.train.knn.Y$crude_birth_rate_perK,
           method="pearson", conf.level=0.95)$estimate # 0.8965095
pearson.BR.test.knn <- 
  cor.test(pred.BR.test.knn, df.BR.test.knn.Y$crude_birth_rate_perK,
           method="pearson", conf.level=0.95)$estimate # 0.9206148
pearson.IM.train.knn <- 
  cor.test(pred.IM.train.knn, df.IM.train.knn.Y$infant_mortality,
           method="pearson", conf.level=0.95)$estimate # 0.8324799
pearson.IM.test.knn <- 
  cor.test(pred.IM.test.knn, df.IM.test.knn.Y$infant_mortality,
           method="pearson", conf.level=0.95)$estimate # .9460087 
pearson.HIV.train.knn <- 
  cor.test(pred.HIV.train.knn, df.HIV.train.knn.Y$HIV_prev,
           method="pearson", conf.level=0.95)$estimate # 0.6945167
pearson.HIV.test.knn <- 
  cor.test(pred.HIV.test.knn, df.HIV.test.knn.Y$HIV_prev,
           method="pearson", conf.level=0.95)$estimate # 0.7394088 
```

### Model Performance

* Birth rate: To compare models, mean square error (MSE) and Pearson coefficients (R) were considered as our model performance metrics. The decision tree had a test MSE higher than that of training, while knn model had a test MSE lower than that of training. Decision tree model resulted in a lower test R, comparing to knn.

* Infant mortality: To compare models, mean square error (MSE) and Pearson coefficients (R) were considered as our model performance metrics. The decision tree had a test MSE higher than that of training, while knn model had a test MSE lower than that of training. The decision tree model resulted in a very small test R, comparing to knn.Knn model performed better than decision tree overall.

* HIV prevalence: regarding this outcome, both models got low results. This is confirmed by results found on the exploratory analysis for HIV prevalence shows that there is no robust linear association with almost none variable.  I could be due these variables are not good to predict this data.

```{r model performance}
# create a data frame to store model performance metrics
model_performance <-
  data.frame(
    outcome = rep(c("birth_rate", "infant_mortality", 
                   "HIV_prevalence"), c(2,2,2)),
    dataset = rep(c("training", "testing"), 3),
    MSE_tree = c(MSE.BR.train.pruned, MSE.BR.test.pruned,
                 MSE.IM.train.pruned, MSE.IM.test.pruned,
                 MSE.HIV.train.pruned, MSE.HIV.test.pruned),
    MSE_knn = c(MSE.BR.train.knn, MSE.BR.test.knn,
                 MSE.IM.train.knn, MSE.IM.test.knn,
                 MSE.HIV.train.knn, MSE.HIV.test.knn),
    Pearson_tree = c(pearson.BR.train.pruned, pearson.BR.test.pruned,
                     pearson.IM.train.pruned, pearson.IM.test.pruned,
                     pearson.HIV.train.pruned, pearson.HIV.test.pruned),
    Pearson_knn = c(pearson.BR.train.knn, pearson.BR.test.knn,
                    pearson.IM.train.knn, pearson.IM.test.knn,
                    pearson.HIV.train.knn, pearson.HIV.test.knn))
model_performance
```

```{r plot pred vs obs, fig.align="center"}
# prepare the data sets for plotting
df.BR.train.tree <- 
  df.BR.train.tree %>% mutate(pred_BR = pred.BR.train.pruned)
df.BR.test.tree <- 
  df.BR.test.tree %>% mutate(pred_BR = pred.BR.test.pruned)
df.BR.train.knn <- 
  df.BR.train.knn %>% mutate(pred_BR = pred.BR.train.knn)
df.BR.test.knn <- 
  df.BR.test.knn %>% mutate(pred_BR = pred.BR.test.knn)
df.IM.train.tree <- 
  df.IM.train.tree %>% mutate(pred_IM = pred.IM.train.pruned)
df.IM.test.tree <- 
  df.IM.test.tree %>% mutate(pred_IM = pred.IM.test.pruned)
df.IM.train.knn <- 
  df.IM.train.knn %>% mutate(pred_IM = pred.IM.train.knn)
df.IM.test.knn <- 
  df.IM.test.knn %>% mutate(pred_IM = pred.IM.test.knn)
df.HIV.train.tree <- 
  df.HIV.train.tree %>% mutate(pred_HIV = pred.HIV.train.pruned)
df.HIV.test.tree <- 
  df.HIV.test.tree %>% mutate(pred_HIV = pred.HIV.test.pruned)
df.HIV.train.knn <- 
  df.HIV.train.knn %>% mutate(pred_HIV = pred.HIV.train.knn)
df.HIV.test.knn <- 
  df.HIV.test.knn %>% mutate(pred_HIV = pred.HIV.test.knn)

# plot predicted values against observed values
df.BR.train.tree %>% 
  ggplot(aes(x = pred_BR, y = crude_birth_rate_perK)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() + theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted Birth Rate") + ylab("Observed Birth Rate") +
  ggtitle("Birth Rate - Decision Tree (Training Set)")
  
df.BR.test.tree %>% 
  ggplot(aes(x = pred_BR, y = crude_birth_rate_perK)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() +  theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted Birth Rate") + ylab("Observed Birth Rate") +
  ggtitle("Birth Rate - Decision Tree (Testing Set)")

df.BR.train.knn %>% 
  ggplot(aes(x = pred_BR, y = crude_birth_rate_perK)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() + theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted Birth Rate") + ylab("Observed Birth Rate") +
  ggtitle("Birth Rate - KNN Regression (Training Set)")

df.BR.test.knn %>% 
  ggplot(aes(x = pred_BR, y = crude_birth_rate_perK)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() +  theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted Birth Rate") + ylab("Observed Birth Rate") +
  ggtitle("Birth Rate - KNN Regression (Testing Set)")

df.IM.train.tree %>% 
  ggplot(aes(x = pred_IM, y = infant_mortality)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() + theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted Infant Mortality") + ylab("Observed Infant Mortality") +
  ggtitle("Infant Mortality - Decision Tree (Training Set)")
  
df.IM.test.tree %>% 
  ggplot(aes(x = pred_IM, y = infant_mortality)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() +  theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted Infant Mortality") + ylab("Observed Infant Mortality") +
  ggtitle("Infant Mortality - Decision Tree (Testing Set)")

df.IM.train.knn %>% 
  ggplot(aes(x = pred_IM, y = infant_mortality)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() + theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted Infant Mortality") + ylab("Observed Infant Mortality") +
  ggtitle("Infant Mortality - KNN Regression (Training Set)")

df.IM.test.knn %>% 
  ggplot(aes(x = pred_IM, y = infant_mortality)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() +  theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted Infant Mortality") + ylab("Observed Infant Mortality") +
  ggtitle("Infant Mortality - KNN Regression (Testing Set)")

df.HIV.train.tree %>% 
  ggplot(aes(x = pred_HIV, y = HIV_prev)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() + theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted HIV Prevalence") + ylab("Observed HIV Prevalence") +
  ggtitle(" HIV Prevalence - Decision Tree (Training Set)")
  
df.HIV.test.tree %>% 
  ggplot(aes(x = pred_HIV, y = HIV_prev)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() +  theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted HIV Prevalence") + ylab("Observed HIV Prevalence") +
  ggtitle(" HIV Prevalence - Decision Tree (Testing Set)")

df.HIV.train.knn %>% 
  ggplot(aes(x = pred_HIV, y = HIV_prev)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() + theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted HIV Prevalence") + ylab("Observed HIV Prevalence") +
  ggtitle(" HIV Prevalence - KNN Regression (Training Set)")

df.HIV.test.knn %>% 
  ggplot(aes(x = pred_HIV, y = HIV_prev)) +
  geom_abline(colour = 'lightgrey') +  
  geom_point() +  theme_bw() + stat_cor(method="pearson") +
  xlab("Predicted HIV Prevalence") + ylab("Observed HIV Prevalence") +
  ggtitle(" HIV Prevalence - KNN Regression (Testing Set)")

```
